init restru
